requirements.txt
amqp==5.1.1
asn1crypto==1.5.1
async-timeout==4.0.2
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
cryptography==36.0.2
Deprecated==1.2.13
Flask==2.1.3
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Moment==1.0.4
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
marshmallow==3.17.0
numpy==1.21.6
oscrypto==1.3.0
packaging==21.3
pandas==1.3.5
prompt-toolkit==3.0.30
psycopg2==2.9.3
pycparser==2.21
pycryptodomex==3.15.0
PyJWT==2.4.0
pymssql==2.2.5
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-json-logger==2.0.4
pytz==2022.1
redis==4.3.4
requests==2.28.1
six==1.16.0
snowflake-connector-python==2.7.9
SQLAlchemy==1.4.39
typing_extensions==4.3.0
urllib3==1.26.10
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.1
